Simplify CLI by removing 'complete' command
- Move prompt handling from complete.go to direct root command
- Refactor code into smaller modules: prompt.go and display.go
- Update README to reflect new simplified usage
- Add Google Gemini models to README
