+++ README.md
@@ -8,7 +8,7 @@ A command-line interface for interacting with Large Language Models (LLMs).
 - Support for specifying different models via flags
 - Configure temperature and system prompts
 - Compare multiple providers side-by-side
-- Support for multiple providers (Anthropic Claude, Deepseek)
+- Support for multiple providers (Anthropic Claude, Deepseek, Google Gemini)
 - Configuration management via config file
 
 ## Installation
@@ -36,6 +36,9 @@ You can set up your API keys in two ways:
    
    # For Deepseek
    gollm set deepseek --api-key your_deepseek_api_key_here
+   
+   # For Google
+   gollm set google --api-key your_google_api_key_here
    ```
 
 2. Using environment variables:
@@ -45,6 +48,9 @@ You can set up your API keys in two ways:
    
    # For Deepseek
    export DEEPSEEK_API_KEY=your_deepseek_api_key_here
+   
+   # For Google
+   export GOOGLE_API_KEY=your_google_api_key_here
    ```
 
 The configuration is stored in `~/.config/gollm/config.yml`.
@@ -53,26 +59,30 @@ The configuration is stored in `~/.config/gollm/config.yml`.
 
 ```bash
 # Get a completion with default model (claude-3-7-sonnet-latest)
-gollm complete "Write a haiku about programming"
+gollm "Write a haiku about programming"
 
 # Using Anthropic's Claude
-gollm complete -m claude-3-7-sonnet-latest "Write a haiku about programming"
+gollm -m claude-3-7-sonnet-latest "Write a haiku about programming"
 
 # Using Deepseek models
-gollm complete -m deepseek-chat "Explain quantum computing"
-gollm complete -m deepseek-coder "Write a function to sort an array in Go"
+gollm -m deepseek-chat "Explain quantum computing"
+gollm -m deepseek-coder "Write a function to sort an array in Go"
+
+# Using Google Gemini models
+gollm -m gemini-2.5-pro-exp-03-25 "Write a function to merge two sorted arrays"
+gollm -m gemini-1.5-flash "Explain parallel computing"
 
 # Adjust the temperature
-gollm complete -t 0.9 "Write a creative story"
+gollm -t 0.9 "Write a creative story"
 
 # Add a system prompt
-gollm complete -s "You are a helpful coding assistant" "How do I implement a binary search in Go?"
+gollm -s "You are a helpful coding assistant" "How do I implement a binary search in Go?"
 
 # Query all configured providers and compare responses
-gollm complete -a "Explain the difference between slices and arrays in Go"
+gollm -a "Explain the difference between slices and arrays in Go"
 
 # Combine flags
-gollm complete -a -t 0.8 -s "You are a Go expert" "What are the best practices for error handling in Go?"
+gollm -a -t 0.8 -s "You are a Go expert" "What are the best practices for error handling in Go?"
 ```
 
 ## Supported Models
@@ -83,6 +93,13 @@ gollm complete -a -t 0.8 -s "You are a Go expert" "What are the best practices f
 ### Deepseek
 - `deepseek-chat`
 - `deepseek-coder`
+
+### Google Gemini
+- `gemini-2.5-pro-exp-03-25`
+- `gemini-2.0-flash`
+- `gemini-2.0-flash-lite`
+- `gemini-1.5-flash`
+- `gemini-1.5-flash-8b`
 
 ## Command-line Options
